+++ inspector-runtime/src/main/java/com/uhufor/inspector/ui/components/SectionTitle.kt
@@ -1,0 +1,26 @@
+package com.uhufor.inspector.ui.components
+
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.text.style.TextOverflow
+import com.uhufor.inspector.ui.compose.dvdp
+import com.uhufor.inspector.ui.compose.dvsp
+
+@Composable
+internal fun SectionTitle(
+    title: String,
+    modifier: Modifier = Modifier,
+) {
+    Text(
+        text = title,
+        modifier = modifier.padding(bottom = 2.dvdp),
+        fontSize = 11.dvsp,
+        fontWeight = FontWeight.Bold,
+        textDecoration = TextDecoration.Underline,
+        overflow = TextOverflow.Ellipsis,
+    )
+}
