+++ inspector-runtime/src/main/java/com/uhufor/inspector/ui/components/SmallButton.kt
@@ -1,0 +1,37 @@
+package com.uhufor.inspector.ui.components
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.uhufor.inspector.R
+import com.uhufor.inspector.ui.compose.dvdp
+import com.uhufor.inspector.ui.compose.dvsp
+
+@Composable
+internal fun SmallButton(
+    text: String,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = modifier
+            .background(
+                color = colorResource(R.color.inspector_button_bg),
+                shape = RoundedCornerShape(4.dp)
+            )
+            .clickable(onClick = onClick)
+            .padding(vertical = 6.dvdp)
+    ) {
+        Text(text = text, fontSize = 9.dvsp, fontWeight = FontWeight.Bold)
+    }
+}
